requirements.txt
# Puppeteer MVP 依赖包
# 截屏和图像处理
mss>=6.1.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Windows输入控制（多种方案）
pyautogui>=0.9.54
pynput>=1.7.6
pygetwindow>=0.0.9
pyscreeze>=0.1.29
pymsgbox>=1.0.9
pytweening>=1.0.4
mouseinfo>=0.1.3
pyperclip>=1.8.2

# Windows系统操作
pywin32>=306
psutil>=5.9.0
wmi>=1.5.1

# 配置管理
PyYAML>=6.0
configparser>=5.3.0

# GUI界面
# tkinter  # 内置模块，无需安装
# PySimpleGUI>=4.60.0  # 需要从私有服务器安装，暂时注释

# 日志和工具
python-json-logger>=2.0.7
colorama>=0.4.6
tqdm>=4.65.0

# 网络和通信
requests>=2.31.0
websockets>=11.0.0

# 数据处理
pandas>=2.0.0
matplotlib>=3.7.0

# 可选：OCR支持
pytesseract>=0.3.10

# 可选：高级图像处理
scikit-image>=0.21.0
scipy>=1.11.0

# 可选：机器学习
scikit-learn>=1.3.0

# 游戏手柄支持
pygame>=2.5.0
inputs>=0.5
xinput>=1.0.0

# 开发和测试
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# 打包和部署
pyinstaller>=5.13.0
cx-Freeze>=6.14.0
